# fix(webapp): disable inline font