--- a/www/webapp/vue.config.js
+++ b/www/webapp/vue.config.js
@@ -4,4 +4,13 @@ module.exports = {
       allowedHosts: 'all',
     },
   },
+  chainWebpack: config => {
+    config.module
+      .rule('fonts')
+      .set('parser', {
+        dataUrlCondition: {
+          maxSize: 0 // Disable inline font to improve FCP (first contentful paint).
+        }
+      });
+  },
 };