Disable inline font to improve FCP (first contentful paint).
Disable to avoid violating CSP rules.

Obsoletes #629