@peterthomassen 2022-08-26
Thanks, - I'm not opposed at all to relax the CSP rules, so perhaps that was a misunderstanding. My point in the other PR was that I did not understand what was going on. Do you know?