@Rotzbua 2022-08-26
@peterthomassen The explanation is a bit longer, so I tried to avoid it ;)
Here https://github.com/desec-io/desec-stack/pull/629#issuecomment-1228620129 is my explanation I hope it is understandable.