@peterthomassen 2022-08-27
ok, let's talk about this. - I'll only get around to take care of it next week, though.