@Rotzbua 2022-09-03
## Performance test

### Setup
* Tested with `vue-cli-service serve --mode production` (gzip http compression enabled by default) 
* Browser Chrome 103

### Results

Lighthouse performance results:

| Metric                   | main    | with pr |
|--------------------------|---------|---------|
| First Contentful   Paint | 5.4 s   | 5.0 s   |
| Time to Interactive      | 5.6 s   | 5.3 s   |
| Speed Index              | 5.4 s   | 5.0 s   |
| Total Blocking Time      | 150 ms  | 160 ms  |
| Largest Contentful Paint | 5.4 s   | 5.0 s   |
| Cumulative Layout Shift  | 0.003   | 0.003   |
|                          |         |         |
| requests                 | 11      | 11      |
| traffic                  | 767 KiB | 696 KiB |