@peterthomassen 2022-10-10
We made some improvements as described in https://github.com/desec-io/desec-stack/issues/631#issuecomment-1273192547, but have not deployed this PR yet.