@peterthomassen 2022-10-10
I'm still torn between whether a tiny wait is a good trade-off for font flickering, and I'd like to understand the issue better.

Here's where I'm lost:
- Your measurements (https://github.com/desec-io/desec-stack/pull/630#issuecomment-1236191388) indicate that disabling font inlining gains about 0.4s to first contentful paint.
- However, even with inline fonts currently enabled, I do see the font being replaced shortly after first contentful paint, so there is flickering *despite* of font inlining.

That indicates that the inline font is actually not used for the first contentful paint. This can be either
- because the CSS is not loaded -- in which case the FCP time should not be affected at all,
- or because the browser doesn't use it for another reason, in which case the question is why the technology exists at all.

Are you able to shed some light on these questions?