@Rotzbua 2022-12-02
> Your measurements (fix(webapp): disable inline font #630 (comment)) indicate that disabling font inlining gains about 0.4s to first contentful paint.

The time reduction comes from the traffic reduction from 767 KiB to 696 KiB.

> However, even with inline fonts currently enabled, I do see the font being replaced shortly after first contentful paint, so there is flickering despite of font inlining.

There is no explicit `font inline` function. It is the generic webpack function: inline tiny files to reduce request to server.

> That indicates that the inline font is actually not used for the first contentful paint.

Yes, you are right.

>  or because the browser doesn't use it for another reason, 

The reason is that not used fonts char-set (greek-ext)  are inlined.

>  in which case the question is why the technology exists at all.

Because webpack makes the decision to inline only by the file size.