@peterthomassen 2022-12-02
> > or because the browser doesn't use it for another reason,
> 
> The reason is that not used fonts char-set (greek-ext) are inlined.

I'm rephrasing what I understood: Even with inlining, there is font flickering in the browser because the inlined font is not used (greek-ext), and the fonts that *are* used are *not* inlined. If that's not correct, please correct me.

If it's correct, we should of course not be inlining the font that is not used (move it to a separate request, or ideally remove it entirely), and instead inline the fonts that are used, to get rid of the flickering. Do you agree with this?