@Rotzbua 2022-12-10
> I'm rephrasing what I understood: Even with inlining, there is font flickering in the browser because the inlined font is not used (greek-ext charset), and the fonts that are used are not inlined. If that's not correct, please correct me.

I think we are talking at cross purposes. The inline does not directly have anything to do with the font flickering.
The inline technique is to save traffic and requests. This is helps for http1.x which has limited concurrent requests.
Also it saves some unnecessary traffic because for file sizes <2kByte the additional http header are about 1kByte. So there is more overhead and content.

The font flicker is caused by the the css property `font-display: swap`.
This enables the browser to start rendering the page as fast as possible. The browser starts rendering the page even if the custom font is not loaded yet. So the webpage is quickly shown with the minimal amount of traffic (good at low bandwidth).
Later if the custom font is loaded the website is rerendered with the custom font. This causes the flicker.

You can find multiple pro con discussions online about this topic.

One solution is to preload the initial shown fonts with the ` <link rel="preload"`. (Not sure how to do this.)
The other solution is to sacrifice the fast rendering and switch back to the default behavior  `font-display: block`. 

To get back to your question: I do not know what impact the inlining of the font has, but the two above solutions are the common and straight forward options.